@@ -13,31 +13,30 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~7.1.4",
-    "@angular/core": "~7.1.4",
-    "@angular/forms": "~7.1.4",
-    "@angular/http": "~7.1.4",
-    "@angular/platform-browser": "~7.1.4",
-    "@angular/platform-browser-dynamic": "~7.1.4",
-    "@angular/router": "~7.1.4",
+    "@angular/common": "^17.0.0",
+    "@angular/core": "^17.0.0",
+    "@angular/forms": "^17.0.0",
+    "@angular/platform-browser": "^17.0.0",
+    "@angular/platform-browser-dynamic": "^17.0.0",
+    "@angular/router": "^17.0.0",
     "@ionic-native/core": "5.0.0-beta.22",
     "@ionic-native/splash-screen": "5.0.0-beta.22",
     "@ionic-native/status-bar": "5.0.0-beta.22",
-    "@ionic/angular": "4.0.0-rc.0",
-    "core-js": "^2.5.7",
-    "rxjs": "6.3.3",
-    "zone.js": "^0.8.26",
+    "@ionic/angular": "^7.0.0",
+    "core-js": "^3.0.0",
+    "rxjs": "^7.8.0",
+    "zone.js": "~0.14.0",
     "moment": "^2.23.0"
   },
   "devDependencies": {
-    "@angular/cli": "~7.1.4",
-    "@angular/compiler": "~7.1.4",
-    "@angular/compiler-cli": "~7.1.4",
-    "@angular/language-service": "~7.1.4",
-    "@angular-devkit/architect": "~0.11.4",
-    "@angular-devkit/build-angular": "~0.11.4",
-    "@angular-devkit/core": "~7.1.4",
-    "@angular-devkit/schematics": "~7.1.4",
+    "@angular/cli": "~17.0.0",
+    "@angular/compiler": "^17.0.0",
+    "@angular/compiler-cli": "^17.0.0",
+    "@angular/language-service": "^17.0.0",
+    "@angular-devkit/architect": "~17.0.0",
+    "@angular-devkit/build-angular": "~17.0.0",
+    "@angular-devkit/core": "~17.0.0",
+    "@angular-devkit/schematics": "~17.0.0",
     "@ionic/angular-toolkit": "^1.2.2",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
@@ -45,15 +44,15 @@
     "codelyzer": "~4.4.2",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.0",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "^2.0.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~10.9.1",
     "tslint": "~5.12.0",
-    "typescript": "~3.1.3"
+    "typescript": "~5.2.0"
   },
   "description": "An Ionic project"
 }
