@@ -15,38 +15,39 @@
     "build:compile-scss": "gulp copy-scss && gulp rename-scss && gulp clean-scss",
     "test": "echo \"Error: no test specified\"",
     "lint": "tslint -c tslint.json 'src/**/*.ts'",
-    "packagr": "ng-packagr -p package.json"
+    "packagr": "ng-packagr -p package.json",
+    "prepare": "npm run build"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/HsuanXyz/ion2-calendar.git"
   },
   "devDependencies": {
-    "@angular/common": "7.1.4",
-    "@angular/compiler": "7.1.4",
-    "@angular/compiler-cli": "^7.2.9",
-    "@angular/core": "7.1.4",
-    "@angular/forms": "7.1.4",
-    "@angular/http": "7.1.4",
-    "@angular/platform-browser": "7.1.4",
-    "@angular/platform-browser-dynamic": "7.1.4",
-    "@angular/router": "7.1.4",
-    "@ionic/angular": "4.0.0-rc.0",
+    "@angular/common": "^17.0.0",
+    "@angular/compiler": "^17.0.0",
+    "@angular/compiler-cli": "^17.0.0",
+    "@angular/core": "^17.0.0",
+    "@angular/forms": "^17.0.0",
+    "@angular/platform-browser": "^17.0.0",
+    "@angular/platform-browser-dynamic": "^17.0.0",
+    "@angular/router": "^17.0.0",
+    "@ionic/angular": "^7.0.0",
     "gulp": "^4.0.0",
     "gulp-bump": "^3.0.0",
     "gulp-clean": "^0.4.0",
     "gulp-ext-replace": "^0.3.0",
-    "gulp-sass": "^4.0.1",
+    "gulp-sass": "^5.1.0",
+    "sass": "^1.77.0",
     "ionicons": "5.1.2",
     "moment": "^2.23.0",
-    "ng-packagr": "^2.0.0",
+    "ng-packagr": "^17.0.0",
     "npm-run-all": "^4.1.2",
-    "rxjs": "6.3.3",
+    "rxjs": "^7.8.0",
     "tsickle": "^0.27.2",
     "tslint": "~5.7.0",
     "tslint-eslint-rules": "^4.1.1",
-    "typescript": "~3.1.3",
-    "zone.js": "0.8.26"
+    "typescript": "~5.2.0",
+    "zone.js": "~0.14.0"
   },
   "keywords": [
     "ionic2",
